python/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-property==0.2.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.3
decorator==5.1.1
deprecation==2.1.0
Django==5.1.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.114.0
fastapi-cli==0.0.5
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.2
hyperframe==5.2.0
idna==2.10
Jinja2==3.1.4
jwcrypto==1.5.6
libretranslatepy==2.1.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
oauthlib==3.2.2
packaging==24.1
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.18.0
pyparsing==3.2.0
python-dotenv==1.0.1
python-keycloak==4.6.2
python-multipart==0.0.9
PyYAML==6.0.2
ratelim==0.1.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.8.1
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.38.5
translate==3.6.1
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.0.1
